requirements.txt
# Install libraries
langchain
openai
PyPDF2
faiss-cpu
tiktoken
python-dotenv
trafilatura